Add Permission on Action

+++ module/Ent/src/Ent/Controller/ContactController.php
@@ -34,6 +34,10 @@ class ContactController extends AbstractActionController
 
     public function listAction()
     {
+        if (!$this->isGranted('list_contact')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $contacts = $this->contactService->getAll();
 
         return new ViewModel(array(
@@ -43,6 +47,10 @@ class ContactController extends AbstractActionController
 
     public function addAction()
     {
+        if (!$this->isGranted('add_contact')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $form = $this->contactForm;
 
         if ($this->request->isPost()) {
@@ -51,7 +59,7 @@ class ContactController extends AbstractActionController
             if ($contact) {
                 $this->flashMessenger()->addSuccessMessage('Le contact a bien été inséré.');
 
-                return $this->redirect()->toRoute('contact');
+                return $this->redirect()->toRoute('zfcadmin/contact');
             }
         }
 
@@ -62,6 +70,10 @@ class ContactController extends AbstractActionController
 
     public function showAction()
     {
+        if (!$this->isGranted('show_contact')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $id = $this->params('id');
 
         $contact = $this->contactService->getById($id);
@@ -77,6 +89,10 @@ class ContactController extends AbstractActionController
 
     public function updateAction()
     {
+        if (!$this->isGranted('update_contact')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $id = $this->params('id');
         $form = $this->contactForm;
         $contact = $this->contactService->getById($id, $form);
@@ -87,7 +103,7 @@ class ContactController extends AbstractActionController
             if ($contact) {
                 $this->flashMessenger()->addSuccessMessage('Le contact a bien été updaté.');
 
-                return $this->redirect()->toRoute('contact');
+                return $this->redirect()->toRoute('zfcadmin/contact');
             }
         }
 
@@ -98,13 +114,17 @@ class ContactController extends AbstractActionController
 
     public function deleteAction()
     {
+        if (!$this->isGranted('delete_contact')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $id = $this->params('id');
 
         $this->contactService->delete($id);
 
         $this->flashMessenger()->addSuccessMessage('Le contact a bien été supprimé.');
 
-        return $this->redirect()->toRoute('contact');
+        return $this->redirect()->toRoute('zfcadmin/contact');
     }
 
 }
